+++ src/Symfony/Component/HttpKernel/RenderingStrategy/RenderingStrategyInterface.php
@@ -1,0 +1,46 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\HttpKernel\RenderingStrategy;
+
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Controller\ControllerReference;
+
+/**
+ * Interface implemented by all rendering strategies.
+ *
+ * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @see Symfony\Component\HttpKernel\HttpContentRenderer
+ */
+interface RenderingStrategyInterface
+{
+    /**
+     * Renders a URI and returns the Response content.
+     *
+     * When the Response is a StreamedResponse, the content is streamed immediately
+     * instead of being returned.
+     *
+     * @param string|ControllerReference $uri     A URI as a string or a ControllerReference instance
+     * @param Request                    $request A Request instance
+     * @param array                      $options An array of options
+     *
+     * @return string|null The Response content or null when the Response is streamed
+     */
+    public function render($uri, Request $request = null, array $options = array());
+
+    /**
+     * Gets the name of the strategy.
+     *
+     * @return string The strategy name
+     */
+    public function getName();
+}