+++ src/Symfony/Component/HttpFoundation/CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 2.2.0
 -----
 
+ * added Request::getTrustedProxies()
+ * deprecated Request::isProxyTrusted()
  * [BC BREAK] JsonResponse does not change a top level array to an object when the array is empty anymore in setData()
  * added a IpUtils class to check if an IP belongs to a CIDR
  * added Request::getRealMethod() to get the "real" HTTP method (getMethod() returns the "intended" HTTP method)